requirements.txt
asgiref==3.8.1
bandit==1.9.3
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-cors-headers==4.7.0
django-csp==4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==25.0.3
hvac==2.4.0
idna==3.10
markdown-it-py==4.0.0
mdurl==0.1.2
packaging==26.0
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.9.0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.3
rich==14.3.2
sqlparse==0.5.3
stevedore==5.6.0
urllib3==2.3.0
whitenoise==6.11.0
